Add files via upload
The TT MAD-2R has FF armor, not implemented in the base game.  HBS under-armored the units with FF and I'm freeing up 1t to add to the stock armor values by 1t. This gives a free weight and armor level closer to the TT.  The HGN-723b has the same problems and HBS granted the unit more free tonnage than TT legal.  I'm using the same solution here for consistency.  In this case, due to TT vs HBS rounding errors, the MAD-2R gets an extra 15 armor, one point of 5 each to the rear torsos to minimize impact.